Ensure session refresh handles missing backend session
