+++ client/src/__mocks__/apiClient.cjs
@@ -1,0 +1,3 @@
+const createAuthClient = jest.fn();
+
+module.exports = { createAuthClient };
